refactor(storage): rename finance to storage
